% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/word_count.R
\name{count_words}
\alias{count_words}
\title{Word Count}
\usage{
count_words(file)
}
\arguments{
\item{file}{A path to a .Rmd file}
}
\description{
Word Count
}
\examples{
# count_words("test.Rmd")
}
\keyword{Word}
\keyword{count}
